+++ libs/core/iostream/tests/unit/direct_adapter.cpp
@@ -57,8 +57,7 @@ void direct_adapter_test()
 
     //--------------indirect_array_ostream------------------------------------//
     {
-        vector<char> dest(
-            static_cast<std::size_t>(data_reps * data_length()), '?');
+        vector<char> dest(data_reps * data_length(), '?');
         indirect_array_ostream out(&dest[0], &dest[0] + dest.size());
         write_data_in_chars(out);
         HPX_TEST_MSG(std::equal(seq.begin(), seq.end(), dest.begin()),
@@ -66,8 +65,7 @@ void direct_adapter_test()
     }
 
     {
-        vector<char> dest(
-            static_cast<std::size_t>(data_reps * data_length()), '?');
+        vector<char> dest(data_reps * data_length(), '?');
         indirect_array_ostream out(&dest[0], &dest[0] + dest.size());
         write_data_in_chunks(out);
         HPX_TEST_MSG(std::equal(seq.begin(), seq.end(), dest.begin()),
@@ -76,16 +74,14 @@ void direct_adapter_test()
 
     //--------------indirect_array_stream-------------------------------------//
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '?');
+        vector<char> test(data_reps * data_length(), '?');
         indirect_array_stream io(&test[0], &test[0] + test.size());
         HPX_TEST_MSG(test_seekable_in_chars(io),
             "failed seeking within indirect_array_stream, in chars");
     }
 
     {
-        vector<char> test(
-            static_cast<std::size_t>(data_reps * data_length()), '?');
+        vector<char> test(data_reps * data_length(), '?');
         indirect_array_stream io(&test[0], &test[0] + test.size());
         HPX_TEST_MSG(test_seekable_in_chunks(io),
             "failed seeking within indirect_array_stream, in chunks");
